Fixing ckeditor5 integration
